create class for convert_dc_line_segments_cim16
